ライブラリが管理するShaderで利用するLightの機能をShaderPartsから分離
