+++ ChCppLibrary/ChDirect3D11Library.h
@@ -13,13 +13,13 @@
 //D3D11AddStruct
 ///////////////////////////////////////////////////////////////////////////////////////
 
-//System操作//
-
 //Object本体//
 #include"Direct3D11/ChTexture/ChTexture11.h"
 #include"Direct3D11/ChMesh/ChMesh11.h"
 #include"Direct3D11/ChPolygonBoard/ChPolygonBoard11.h"
 #include"Direct3D11/ChSprite/ChSprite11.h"
+//ライブラリで用意したShaderを利用するクラス//
+#include"Direct3D11/ChShaderParts/ChLight/ChLight11.h"
 //空間操作//
 #include"Direct3D11/ChCameraController/ChCameraController11.h"
 //Object操作//
